fix(statement): move SQL-emission helpers out of diff.go
quoteIdent, quoteIdentList, and formatPartitionValue are SQL-emission
helpers, not Diff-specific logic, and diff.go is already 1600+ lines.
Move them to a new pkg/statement/utils.go alongside their tests in
utils_test.go. All functions stay package-private (lowercase) and the
imports in diff.go are unchanged — strconv / strings / sqlescape are
still used by other code there.

No behaviour change.
